pyproject.toml
[tool.poetry]
name = "forecast"
version = "0.1.0"
description = ""
authors = ["Yevhenii Konyshev <evgeniy.konyshev2007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
aiohttp = "^3.9.1"
pandas = "^2.2.0"
numpy = "^1.26.3"
pydantic = "^2.5.3"
pydantic-extra-types = "^2.4.1"
uvloop = "^0.19.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
isort = "*"
pyright = "^1"
mypy = "^1"
ruff = "*"
flake8 = "^7.0.0"
pytest-asyncio = "^0.23.3"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"

[tool.isort]
profile = "black"

[tool.ruff]
select = ['E', 'F', 'W', 'I', 'N', 'UP', 'E', 'W']
ignore = []

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
typeCheckingMode = "strict"
useLibraryCodeForTypes = true

reportMissingImports = true
reportMissingTypeStubs = false

reportUnknownMemberType = false
reportUnknownVariableType = false

strictListInference = true
stringDictionaryInference = true
strictSetInference = true
reportUnusedImport = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
